+++ src/DynamicalBilliards.jl
@@ -10,7 +10,7 @@ fast implementation of billiard systems of arbitrary construction.
 module DynamicalBilliards
 
 using StaticArrays
-import Base.show
+import Base: show, eltype, getindex
 
 const SV = SVector{2}
