+++ src/billiards/standard_billiards.jl
